Done article view CSS

+++ public/react/components/ArticleView.js
@@ -4,6 +4,10 @@ import apiURL from '../api';
 export const ArticleView = ({page, goback, setEditorMode, fetchPages}) => {
     // The API returns everything there's absolutely no point making more requests
     async function deleteArticle() {
+        const doubleCheck = confirm("You are about to delete '" + page.title + "', are you sure?");
+        if (doubleCheck == false) {
+            return; // Cancel if they say no
+        }
         const response = await fetch(apiURL + '/wiki/' + page.slug, {
             method: "DELETE"
         });
@@ -13,14 +17,27 @@ export const ArticleView = ({page, goback, setEditorMode, fetchPages}) => {
         goback();
     }
     return (<>
-    <span onClick={goback} className="return-button">Go back</span>
-    <h2>{page.title}</h2>
+    <div className="split-wrapping">
+        <div className="split-box">
+            <h2>{page.title}</h2>
+        </div>
+        <div className="split-box right-lean">
+            <span onClick={goback}>Go back</span>
+        </div>
+    </div>
     <div className="article-options">
         <span><b>View</b> | <u onClick={() => setEditorMode(true)}>Edit</u></span>
     </div>
     <article>
         {page.content}
     </article>
-    <button onClick={deleteArticle}>Delete</button>
+    <div className="split-wrapping">
+        <div className="split-box">
+            <p>Data</p>
+        </div>
+        <div className="split-box right-lean">
+            <button onClick={deleteArticle}>Delete</button>
+        </div>
+    </div>
     </>)
 }
